V1, add selective anononymisation

+++ anonymise.py
@@ -15,12 +15,15 @@ FILES = [
     '.zip'
 ]
 
+global name_counter
+
 def get_size(filename):
     file_stats = os.stat(filename)
     return file_stats.st_size
 
 def update_counter():
-    name_counter += 1
+    global name_counter
+    name_counter = str(int(name_counter) + 1)
 
 def get_files_to_rename():
 
@@ -33,24 +36,30 @@ def get_files_to_rename():
     for file_name in all_files:
         if "-" in file_name and ".zip" in file_name:
             files_to_rename.append(
-                fine_name.split("_")[0]
+                file_name.split("_")[0]
                 + "_" +
-                fine_name.split("_")[1]
+                file_name.split("_")[1]
             )
-    
+
     return files_to_rename
 
 def rename_file(src, dest):
     '''
     Rename a single file and check it was successful.
     '''
-    size_before = get_size(src)
 
-    assert(not os.path.exists(dest))
-    os.rename(src, dest)
+    if (os.path.exists(src)):
 
-    size_after = get_size(dest)
-    assert(size_before == size_after)
+        size_before = get_size(src)
+
+        assert(not os.path.exists(dest))
+        os.rename(src, dest)
+
+        size_after = get_size(dest)
+        assert(size_before == size_after)
+
+    else:
+        print(src + " does not exist.")
 
     print(dest + " renamed.")
 
@@ -64,15 +73,13 @@ def rename_all(files_to_rename):
             src_base = os.path.join(cwd, base_name)
             src_base_exp = src_base + expression
             
-            dest_base = os.path.join(cwd, str(name_counter))
+            dest_base = os.path.join(cwd, name_counter)
             dest_base_exp = dest_base + expression
 
             for file_name in FILES:
                 src = src_base_exp + file_name
                 dest = dest_base_exp + file_name
-                print(src)
-                print(dest)
-                #rename_file(src, dest)
+                rename_file(src, dest)
 
         update_counter()
 
@@ -81,7 +88,7 @@ parser = ArgumentParser()
 parser.add_argument("-n", "--starting_number", required=True, help="The number name to start")
 args = parser.parse_args()
 
-global name_counter = args.starting_number
+name_counter = args.starting_number
 
 files_to_rename = get_files_to_rename()
 rename_all(files_to_rename)
